Refactor Datasets and Notebooks context to include rightBarContent state management and update related components for improved functionality
